+++ erkc63/payment.py
@@ -1,5 +1,6 @@
 import dataclasses as dc
 import datetime as dt
+from decimal import Decimal
 
 
 @dc.dataclass(frozen=True)
@@ -12,7 +13,7 @@ class Payment:
 
     date: dt.date
     """Дата"""
-    summa: float
+    summa: Decimal
     """Сумма"""
     provider: str
     """Платежный провайдер"""
